Ensure metadata reflects aggregated scores.
When we aggregate multiple trans scores we end up with a state where
individual scoreids in metadata no longer point to correct columns in
the scores matrix. This is resolved by creating new scoreids both in
metadata and in scores matrix. These are named as: X_gwasid_trans and
will be repeated in the metadata for each score which was aggregated
in the scores matrix.

This fix also adds a few checks to be absolutely sure that we do not end
up in a position when there are some scoreids in metadata which point to
non-existing scores in scores matrix.